// src/app/model/cars/cars.ts
export interface Cars {
  id: number;
  title: string;
  adress: string;
  brand: string;
  model: string;
  model_year: number;
  issuance: Date;
  mileage: number;
  fuel: string;
  color: string;
  numbers_doors: number;
  horse_power: number;
  price: number;
  pictures:string[];
  solde: boolean;
}